working on encoder

+++ pkg/zkc/vm/internal/bytecode/bytecode.go
@@ -25,65 +25,97 @@ package bytecode
 // Currently, n is instruction specific.
 const (
 	// FAIL instruction
-	FAIL = uint8(0)
-	// JMP (jump unconditional) instruction
-	JMP = uint8(1)
-	// JIF (jump conditional) instruction
-	JIF = uint8(2)
+	FAIL = uint32(0)
+	// JMP (jump unconditional) instruction.  Format of this instruction is:
+	//
+	//	31                       8 7    5 4       0
+	// +--------------------------+------+---------+
+	// |        offset            | n/a  | opcode  |
+	// +--------------------------+------+---------+
+	//
+	// Here, offset is a signed u16 relative offset, where the following
+	// instruction is considered to be at offset 0.
+	JMP = uint32(1)
+	// JIF (jump conditional) instruction.  Format of this instruction is:
+	//
+	//	31                                0
+	// +--------+--------+--------+------+--------+
+	// | offset |  rs0   |  rs1   |  op  | opcode |
+	// +--------+--------+--------+------+--------+
+	//
+	// Here, offset is a signed u8 relative offset, where the following
+	// instruction is considered to be at offset 0.  Likewise, rs0 and rs1 are
+	// u8 source registers, whilst op identifies the operation.
+	JIF = uint32(2)
 	// CALL instruction
-	CALL = uint8(3)
+	CALL = uint32(3)
 	// RET instruction
-	RET = uint8(4)
+	RET = uint32(4)
 	// LOAD instruction
-	LOAD = uint8(5)
+	LOAD = uint32(5)
 	// STORE instruction
-	STORE = uint8(6)
+	STORE = uint32(6)
 	// PUSH instruction
-	PUSH = uint8(7)
+	PUSH = uint32(7)
 	// POP instruction
-	POP = uint8(8)
-	// MOVE instruction
-	MOVE = uint8(9)
+	POP = uint32(8)
+	// MOVE instruction.  Format of this instruction is:
+	//
+	//  31                                       0
+	// +--------+--------+--------+------+--------+
+	// |   n/a  |   rs   |   rd   | n/a  | opcode |
+	// +--------+--------+--------+------+--------+
+	//
+	// Here, rs is a u8 source register whilst rd is a u8 destination register.
+	MOVE = uint32(9)
 	// DESTRUCT instruction
-	DESTRUCT = uint8(10)
+	DESTRUCT = uint32(10)
 	// CAST instruction
-	CAST = uint8(11)
-	// ADD instruction
-	ADD = uint8(12)
+	CAST = uint32(11)
+	// ADD instruction.  Format of this instruction is:
+	//
+	//  31                                       0
+	// +--------+--------+--------+------+--------+
+	// |  rs0   |  rs1   |   rd   | n/a  | opcode |
+	// +--------+--------+--------+------+--------+
+	//
+	// Here, rs0 and rs1 are u8 source registers, whilst rd is a u8 destination
+	// register.
+	ADD = uint32(12)
 	// ADDC (add with constant) instruction
-	ADDC = uint8(13)
+	ADDC = uint32(13)
 	// SUB instruction
-	SUB = uint8(14)
+	SUB = uint32(14)
 	// SUBC (subtract with constant) instruction
-	SUBC = uint8(15)
+	SUBC = uint32(15)
 	// CSUB (subtract from constant) instruction
-	CSUB = uint8(16)
+	CSUB = uint32(16)
 	// MUL instruction
-	MUL = uint8(17)
+	MUL = uint32(17)
 	// MULC (multiply with constant) instruction
-	MULC = uint8(18)
+	MULC = uint32(18)
 	// DIV instruction
-	DIV = uint8(19)
+	DIV = uint32(19)
 	// ADDMOD_P instruction
-	ADDMOD_P = uint8(20)
+	ADDMOD_P = uint32(20)
 	// SUBMOD_P instruction
-	SUBMOD_P = uint8(21)
+	SUBMOD_P = uint32(21)
 	// MULMOD_P instruction
-	MULMOD_P = uint8(22)
+	MULMOD_P = uint32(22)
 	// AND instruction
-	AND = uint8(23)
+	AND = uint32(23)
 	// OR instruction
-	OR = uint8(24)
+	OR = uint32(24)
 	// XOR instruction
-	XOR = uint8(25)
+	XOR = uint32(25)
 	// NOT instruction
-	NOT = uint8(26)
+	NOT = uint32(26)
 	// SHL instruction
-	SHL = uint8(27)
+	SHL = uint32(27)
 	// SHR instruction
-	SHR = uint8(28)
+	SHR = uint32(28)
 	// CAT instruction
-	CAT = uint8(29)
+	CAT = uint32(29)
 	//
 )
